Separate Axis plugins now are grouped in a single file
