+++ src/main/resources/plugins.config
@@ -19,14 +19,14 @@
 
 Plugins>NMQC, "Planar Uniformity", NMQC.Planar_Uniformity
 Plugins>NMQC, "Tomographic Uniformity", NMQC.Tomographic_Uniformity
-Plugins>NMQC, "Four Bar Phantom", NMQC.Four_Bar_Phantom
-Plugins>NMQC, "Quadrant Bar Phantom", NMQC.Bar_Quadrant
 Plugins>NMQC>Center of Rotation, "Sine Fit", NMQC.C_O_R("Sine")
 Plugins>NMQC>Center of Rotation, "Conjugate Views", NMQC.C_O_R("Conjugate")
-Plugins>NMQC>Intrinsic Resolution and Linearity, "IRL X", NMQC.IntResol_Linearity_X
-Plugins>NMQC>Intrinsic Resolution and Linearity, "IRL Y", NMQC.IntResol_Linearity_Y
-Plugins>NMQC>Spatial Resolution and Pixel_Size, "SSR PxSz X", NMQC.SSR_PxSz_X
-Plugins>NMQC>Spatial Resolution and Pixel_Size, "SSR PxSz Y", NMQC.SSR_PxSz_Y
+Plugins>NMQC>Intrinsic Resolution and Linearity, "Horizontal Stripes", NMQC.IntResol_Linearity("Horizontal")
+Plugins>NMQC>Intrinsic Resolution and Linearity, "Vertical Stripes", NMQC.IntResol_Linearity("Vertical")
+Plugins>NMQC>Spatial Resolution and Pixel_Size, "Vertical Bars", NMQC.SSR_PxSz("Vertical")
+Plugins>NMQC>Spatial Resolution and Pixel_Size, "Horizontal Bars", NMQC.SSR_PxSz("Horizontal")
+Plugins>NMQC>Spatial Resolution and Pixel_Size, "Four Bars", NMQC.Four_Bar_Phantom
+Plugins>NMQC>Spatial Resolution and Pixel_Size, "Quadrant Bar", NMQC.Bar_Quadrant
 Plugins>NMQC>Tools, "Geometric Mean", NMQC.Tools.Geometric_Mean
 Plugins>NMQC>Tools, "Cardiac Reslicer", NMQC.Tools.Cardiac_Reslicer
